Working cub3d

+++ includes/cub3d.h
@@ -6,7 +6,7 @@
 /*   By: yelaissa <yelaissa@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/08/02 18:07:29 by yelaissa          #+#    #+#             */
-/*   Updated: 2023/08/11 18:54:50 by yelaissa         ###   ########.fr       */
+/*   Updated: 2023/08/12 11:19:25 by yelaissa         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -54,6 +54,6 @@ void	cast_ray();
 int		get_rgba(int r, int g, int b, float a);
 float	normalize_angle(float rayAngle);
 t_ray	get_ray(float ray_angle);
-int 	distance_between_points(float x1, float y1, float x2, float y2);
+float 	distance_between_points(float x1, float y1, float x2, float y2);
 void	rect(float x, float y, int width, int height, int color);
 #endif
